finish baby

+++ ventas/src/cliente/index.js
@@ -4,6 +4,7 @@ import ClientDataServices from '../services/ClientDataServices'
 import Modal from "react-bootstrap/Modal";
 import Button from 'react-bootstrap/Button';
 import AddClient from '../AddClient';
+import Swal from 'sweetalert2'
 
 const Cliente = () => {
 
@@ -14,14 +15,20 @@ const Cliente = () => {
     const [apellido, setApellido] = useState("");
     const [celular, setCelular] = useState("");
     const [id, setId] = useState("");
+    const [nit, setNit] = useState("");
 
 
-    const handleModal = (idx, name, last, cel) => {
+    const handleModal = (idx, name, last, cel, nit) => {
         setShow(!show);
         setId(idx);
         setNombre(name);
         setApellido(last);
         setCelular(cel);
+        setNit(nit);
+    }
+
+    const modalXx = () => {
+        setShow(!show)
     }
 
     useEffect(() => {
@@ -35,8 +42,25 @@ const Cliente = () => {
     }
 
     const deleteHandler = async (id) => {
-        await ClientDataServices.deleteClient(id);
-        getClients();
+        await Swal.fire({
+            title: 'Eliminar',
+            text: "Estás seguro que deseas eliminar este producto?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, Eliminar'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                ClientDataServices.deleteClient(id);
+                getClients();
+              Swal.fire(
+                'Eliminado!',
+                'El producto se ha eliminado!',
+                'success'
+              )
+            }
+          })
     }
 
     return (
@@ -49,6 +73,9 @@ const Cliente = () => {
                             {
                                 Clients.map(item => (
                                     <li className="list-group-item" key={item.id}>
+                                        <strong>Nit: </strong>
+                                        <span>{item.nit}</span>
+                                        <br />
                                         <strong>Nombre: </strong>
                                         <span>{item.nombre}</span>
                                         {' '}
@@ -63,7 +90,7 @@ const Cliente = () => {
                                         </button>
                                         <Button
                                             className="btn btn-warning btn-sm float-right mr-2"
-                                            onClick={() => handleModal(item.id, item.nombre, item.apellido, item.celular)}
+                                            onClick={() => handleModal(item.id, item.nombre, item.apellido, item.celular, item.nit)}
                                         >
                                             Editar
                                         </Button>
@@ -74,7 +101,7 @@ const Cliente = () => {
                         </ul>
                     </div>
                     <div className="col-md-6">
-                        <AddClient />
+                        <AddClient GetClients={getClients} />
                     </div>
                 </div>
             </div>
@@ -83,7 +110,7 @@ const Cliente = () => {
                     <Modal.Title>Editar Cliente</Modal.Title>
                 </Modal.Header>
                 <Modal.Body>
-                    <EditClient Id={id} Nombre={nombre} Apellido={apellido} Celular={celular} />
+                    <EditClient Id={id} Nombre={nombre} Apellido={apellido} Celular={celular} Nit={nit} GetClients={getClients} ModalXx={modalXx} />
                 </Modal.Body>
             </Modal>
         </div>
